Lint cleanup (repo-wide, `make lint` now zero warnings) + HIL jitter alert split out
1. Pre-existing lint findings, none of them caused by the typesupport work, just never cleaned
   up:
   - examples/linux/{uint64/{publisher,subscriber},set_bool/{client,server},ping_pong/{ping,pong}}.c:
     removed unused <stdlib.h>/<string.h> (neither ever called anything from either - same
     "misc-include-cleaner" finding perf_client.c/perf_server.c already had fixed for the same
     reason during the Bulk cutover).
   - examples/linux/common/cli_opts.h: `1u` -> `1U` (readability-uppercase-literal-suffix, 5
     spots). cli_opts.c: added a direct <stdint.h> for uint32_t (previously only reached
     transitively through cli_opts.h).
   - src/hal_linux.c: tt_send_iov()'s `(void*)(uintptr_t)hdr`/`body` casts were a redundant
     integer round-trip just to discard `const` on an already-a-pointer value (performance-no-
     int-to-ptr) - a straight `(void*)hdr` does the same thing without it, same fix already
     applied to the generated Bulk.c during the earlier cutover. Also added NOLINT(misc-include-
     cleaner) on <sys/uio.h> and its own `struct iovec` use - clang-tidy's IWYU mapping doesn't
     know this glibc symbol's real (portable, POSIX-specified) home and flags both sides of a
     correct, portable include; same class of false positive this file already suppresses for
     CLOCK_REALTIME/SOL_SOCKET a few lines up.

2. performance.yml: rtt mdev (ping's own mean-deviation/jitter stat) is inherently noisy on real
   hardware - it alone tripped the Performance Test's 200% alert-threshold three separate times
   in one session, each time on a commit nowhere near the ping/pong latency path (a check-all.yml
   fix, a doc-only CHANGELOG commit, ...). Split into its own benchmark group (run_perf.sh now
   writes a separate latency-jitter-benchmark.json) with fail-on-alert: false - still tracked and
   graphed (summary-always, auto-push, same as every other group), just never fails the build on
   its own. rtt avg and packet loss - the two latency-benchmark.json still carries - stay at
   200%/fail-on-alert: true, since neither showed this flakiness and both do reflect a real
   regression when they move.

Verified: `make lint` / `make -C platform/freertos lint` both zero warnings (previously ~16
pre-existing findings across 9 files); `make test` / `make sanitize` still pass; run_perf.sh's
new JSON-writing logic checked directly (valid JSON, correct split) since HIL itself isn't
reachable from here to run the real script end to end.

+++ src/hal_linux.c
@@ -20,7 +20,11 @@
 #include <arpa/inet.h>
 #include <netinet/in.h>
 #include <sys/socket.h>
-#include <sys/uio.h>
+// struct iovec (tt_send_iov, below) - clang-tidy's IWYU mapping doesn't know this glibc symbol's
+// real (portable, POSIX-specified) home, so it flags both this include and the struct itself as
+// if neither provided/needed the other; this is the correct, portable header regardless (see
+// readv(2)/writev(2)/sendmsg(2)).
+#include <sys/uio.h> // NOLINT(misc-include-cleaner)
 #include <tickle/config.h>
 #include <tickle/hal.h>
 #include <tickle/tickle.h>
@@ -159,9 +163,10 @@ int32_t tt_send_to(struct tt_Node* node, const void* buf, size_t len, uint32_t i
 
 int32_t tt_send_iov(struct tt_Node* node, const void* hdr, size_t hdr_len, const void* body, size_t body_len,
                     uint32_t ip, uint16_t port) {
+    // NOLINTNEXTLINE(misc-include-cleaner) - see <sys/uio.h>'s own include comment
     struct iovec iov[2] = {
-        {.iov_base = (void*)(uintptr_t)hdr, .iov_len = hdr_len},
-        {.iov_base = (void*)(uintptr_t)body, .iov_len = body_len},
+        {.iov_base = (void*)hdr, .iov_len = hdr_len},
+        {.iov_base = (void*)body, .iov_len = body_len},
     };
 
     struct sockaddr_in unicast_addr;
